conf.py -> conftest.pz
