+++ src/kartograf/tests/test_utils_vis.py
@@ -12,7 +12,7 @@ try:
 except ImportError:
     optional_imports = False
 
-from .conf import stereo_chem_mapping
+from .conftest import stereo_chem_mapping
 
 
 if(optional_imports):
